[REFACTOR] Changed medical-equipment tests to import wrapper dashboard files instead of modules
